Refactor out the chunking code
